+++ index.js
@@ -2,7 +2,7 @@ const express = require('express')
 const request = require('request')
 const Blockchain = require('./blockchain')
 const bodyParser = require('body-parser')
-const PubSub = require('./pubsub')
+const PubSub = require('./app/pubsub')
 
 const app = express()
 const blockchain = new Blockchain()
